+++ src/leg.rs
@@ -45,11 +45,11 @@ impl Leg {
         let accums = self.accumulate_for_leg(&env.sens_values);
         let activs = self.activate(&accums);
         self.update_values(&activs);
-        env.act(activs[0], activs[1]);
+        env.act(activs[0] as usize, activs[1] as usize);
     }
 
     fn accumulate_for_leg(&self, sens_values: &[i16; 3]) -> Vec<i16> {
-        let values: Vec<usize> = self.0.iter().map(|neuron| neuron.value).collect();
+        let values: Vec<i16> = self.0.iter().map(|neuron| neuron.value).collect();
         let out_accums: Vec<i16> = self
             .0
             .iter()
@@ -64,7 +64,7 @@ impl Leg {
     }
 
     pub fn accumulate_for_bot(&self, sens_value_lists: &[[i16; 3]], leg_index: usize) -> Vec<i16> {
-        let values: Vec<usize> = self.0.iter().map(|neuron| neuron.value).collect();
+        let values: Vec<i16> = self.0.iter().map(|neuron| neuron.value).collect();
         let out_sens_values = sens_value_lists[leg_index];
         let out_accums: Vec<i16> = self
             .0
@@ -83,7 +83,7 @@ impl Leg {
         [out_accums, hid_accums].concat()
     }
 
-    pub fn activate(&self, accs: &[i16]) -> Vec<usize> {
+    pub fn activate(&self, accs: &[i16]) -> Vec<i16> {
         self.0
             .iter()
             .zip(accs.iter())
@@ -91,7 +91,7 @@ impl Leg {
             .collect()
     }
 
-    pub fn update_values(&mut self, values: &[usize]) {
+    pub fn update_values(&mut self, values: &[i16]) {
         self.0
             .iter_mut()
             .zip(values.iter())
